% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DonneesTest.R
\docType{data}
\name{DonneesTest}
\alias{DonneesTest}
\title{Les données nettoyées de l'essence}
\format{
Un tableur aec 9995 lignes et 10 colonnes
}
\usage{
data(DonneesTest)
}
\description{
stations et prix de l'essence
}
\details{
\itemize{
\item id. id de la station
\item pop. R si route, A si autoroute
\item stationName. nom et adresse
\item latitude. numeric
\item longitude. numeric
\item departement. code dep en caracteres
\item marque. enseigne renseignée de la station
\item Gazole. prix. caracteres
\item E10 prix. caracteres
\item SP98 prix. caracteres
}
}
\keyword{datasets}
