% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importandcleandata.R
\name{ImportAndCleanData}
\alias{ImportAndCleanData}
\title{ImportAndCleanData : Recupere les Prix}
\usage{
ImportAndCleanData(urlgdoc, urlgsheet)
}
\arguments{
\item{urlgdoc}{: l'url d'un google doc provenant du site gouvernemental}

\item{urlgsheet}{: le titre de la feuille contenant les données}
}
\value{
un tibble contenant les prix du gazole, du SP98 et du E10
}
\description{
ImportAndCleanData : Recupere les Prix
}
\examples{
urlgdoctest<-"https://docs.google.com/spreadsheets/d/1aTGGqcUAjWg8Lz5sqQ5CzCEN7LbZdKGV_PdQC4YZ6Qw/edit?usp=sharing"
urlgsheettest<-"26-janv"
# you may need to relog to avoid the "Request failed [429]. Retry " error, so use
# gs4_auth() 
DonneesTest<-ImportAndCleanData(urlgdoc = urlgdoctest, urlgsheet = urlgsheettest)
head(DonneesTest)
}
