% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sortir-les-prix-moyens-et-ma-c-dians-getprices.R
\name{GetPrices}
\alias{GetPrices}
\title{GetPrices}
\usage{
GetPrices(dataset, fonction = "mean", dobydepartement = TRUE)
}
\arguments{
\item{dataset}{tibble un jeu de données propre}

\item{fonction}{chaine de caracteres mean ou median selon la métrique souhaite}

\item{dobydepartement}{logical selon qu'on veut le resultat par departement ou agrégé}
}
\value{
un tibble contenant les prix moyens ou médians du gazole, du SP98 et du E10
}
\description{
GetPrices
}
